Refactoering of the whole procejt
