--- src/main/java/app/enums/continents.java
@@ -1,5 +1,0 @@
-package app.enums;
-
-public enum Continents {
-ASIA, AFRICA, NORTH_AMERICA, SOUTH_AMERICA, ANTARCTICA, EUROPE, AUSTRALIA
-}
