+++ src/main/java/app/routes/SightRoute.java
@@ -9,22 +9,24 @@ import jakarta.persistence.EntityManagerFactory;
 import static io.javalin.apibuilder.ApiBuilder.*;
 
 public class SightRoute {
-    private EntityManagerFactory emf = HibernateConfig.getEntityManagerFactory("country");
-    private SightDAO sightDAO = new SightDAO(emf);
-    private SightController sightController = new SightController(sightDAO);
+    private SightController sightController;
+
+    public SightRoute(EntityManagerFactory emf) {
+        sightController = new SightController(new SightDAO(emf));
+    }
 
     protected EndpointGroup addSightRoutes() {
         return () -> {
             //GET ALL ROOMS
-            get("/sight", sightController::getAll);
+            get("/", sightController::getAll);
 
-            get("/sight/{id}", sightController::getById);
+            get("/{id}", sightController::getById);
 
-            post("/sight", sightController::create);
+            post("/", sightController::create);
 
-            put("/sight", sightController::update);
+            put("/{id}", sightController::update);
 
-            delete("/sight/{id}", sightController::delete);
+            delete("/{id}", sightController::delete);
 
         };
     }
